chore(RHINENG-16141): Cleanup EditPolicyProfileRules & TabbedRules
