--- src/PresentationalComponents/TabbedRules/index.js
@@ -1,6 +1,0 @@
-export { default as TabbedRules } from './TabbedRules';
-export {
-  profilesWithRulesToSelection,
-  tabsDataToOsMinorMap,
-  extendProfilesByOsMinor,
-} from './ruleSelection';
