+++ src/PresentationalComponents/Tailorings/osVersionText.js
