+++ src/SmartComponents/CreatePolicy/EditPolicyProfilesRules/EditPolicyProfilesRules.js
@@ -1,34 +1,215 @@
-import React from 'react';
+import React, { useCallback } from 'react';
 import { compose } from 'redux';
 import { connect } from 'react-redux';
-import { Bullseye } from '@patternfly/react-core';
-import Spinner from '@redhat-cloud-services/frontend-components/Spinner';
+import {
+  Bullseye,
+  EmptyState,
+  EmptyStateBody,
+  EmptyStateHeader,
+  Spinner,
+  Text,
+  TextContent,
+  TextVariants,
+} from '@patternfly/react-core';
 import propTypes from 'prop-types';
 import {
   formValueSelector,
   reduxForm,
   propTypes as reduxFormPropTypes,
 } from 'redux-form';
-import useAPIV2FeatureFlag from '../../../Utilities/hooks/useAPIV2FeatureFlag';
-import EditPolicyProfilesRulesRest from './EditPolicyProfilesRulesRest';
-import EditPolicyProfilesRulesGraphQL from './EditPolicyProfilesRulesGraphQL';
+import { useDeepCompareEffectNoCheck } from 'use-deep-compare-effect';
+import {
+  StateViewPart,
+  StateViewWithError,
+  Tailorings,
+} from '@/PresentationalComponents';
+import * as Columns from '@/PresentationalComponents/RulesTable/Columns';
+import useProfileRuleIds from './useProfileRuleIds';
 
-const EditPolicyProfilesRules = (props) => {
-  const apiV2Enabled = useAPIV2FeatureFlag();
+const EditPolicyProfilesRules = ({
+  profile,
+  selectedRuleRefIds,
+  change,
+  osMajorVersion,
+  osMinorVersionCounts,
+  valueOverrides = {},
+}) => {
+  const preselected =
+    selectedRuleRefIds &&
+    (selectedRuleRefIds || []).reduce(
+      (prev, cur) => ({
+        ...prev,
+        [cur.osMinorVersion]: cur.ruleRefIds,
+      }),
+      {}
+    );
+  const profileRefId = profile?.ref_id;
 
-  return apiV2Enabled === undefined ? (
+  const skipFetchingProfileRuleIds =
+    !osMajorVersion ||
+    !profileRefId ||
+    (osMinorVersionCounts || []).length === 0;
+
+  const {
+    profilesAndRuleIds,
+    loading: preselectedRuleIdsLoading,
+    error: preselectedRuleIdsError,
+  } = useProfileRuleIds({
+    profileRefId,
+    osMajorVersion,
+    osMinorVersions: osMinorVersionCounts.map(
+      ({ osMinorVersion }) => osMinorVersion
+    ),
+    skip: skipFetchingProfileRuleIds,
+  });
+
+  const onSelect = useCallback(
+    (
+      _securityGuideId,
+      { os_minor_version: osMinorVersion },
+      newSelectedRuleIds
+    ) => {
+      const updatedSelectedRuleRefIds = structuredClone(
+        selectedRuleRefIds || []
+      );
+
+      if (updatedSelectedRuleRefIds.length === 0) {
+        if (newSelectedRuleIds.length === 0) {
+          return;
+        }
+
+        updatedSelectedRuleRefIds.push({
+          osMinorVersion,
+          ruleRefIds: newSelectedRuleIds,
+        });
+      } else {
+        const index = updatedSelectedRuleRefIds.findIndex(
+          ({ osMinorVersion: _osMinorVersion }) =>
+            _osMinorVersion === osMinorVersion
+        );
+
+        updatedSelectedRuleRefIds[index].ruleRefIds = newSelectedRuleIds;
+      }
+      change('selectedRuleRefIds', updatedSelectedRuleRefIds);
+    },
+    [change, selectedRuleRefIds]
+  );
+
+  useDeepCompareEffectNoCheck(() => {
+    if (profilesAndRuleIds !== undefined && selectedRuleRefIds === undefined) {
+      change(
+        'selectedRuleRefIds',
+        osMinorVersionCounts.map(({ osMinorVersion }) => ({
+          osMinorVersion,
+          ruleRefIds: profilesAndRuleIds.find(
+            ({ osMinorVersion: profileOsMinorVersion }) =>
+              profileOsMinorVersion === osMinorVersion
+          ).ruleIds,
+        }))
+      );
+    }
+  }, [change, osMinorVersionCounts, profilesAndRuleIds, selectedRuleRefIds]);
+
+  const onValueOverrideSave = (
+    _policy,
+    osMinorVersion,
+    valueDefinition,
+    valueOverrideValue,
+    closeInlineEdit
+  ) => {
+    let valueOverridesUpdated = structuredClone(valueOverrides);
+
+    if (valueOverrideValue === '' || valueOverrideValue === undefined) {
+      delete valueOverridesUpdated?.[osMinorVersion]?.[valueDefinition.id];
+    } else {
+      valueOverridesUpdated = {
+        ...valueOverrides,
+        [osMinorVersion]: {
+          ...valueOverrides[osMinorVersion],
+          [valueDefinition.id]: valueOverrideValue,
+        },
+      };
+    }
+
+    change('valueOverrides', valueOverridesUpdated);
+    closeInlineEdit();
+  };
+
+  const noRuleSets =
+    !preselectedRuleIdsError &&
+    !preselectedRuleIdsLoading &&
+    Object.keys(profilesAndRuleIds || {}).length === 0;
+
+  return !preselected ? (
     <Bullseye>
       <Spinner />
     </Bullseye>
-  ) : apiV2Enabled ? (
-    <EditPolicyProfilesRulesRest {...props} />
   ) : (
-    <EditPolicyProfilesRulesGraphQL {...props} />
+    <React.Fragment>
+      <TextContent className="pf-v5-u-pb-md">
+        <Text component={TextVariants.h1}>Rules</Text>
+        <Text>
+          Customize your <b>{profile.title}</b> SCAP policy by including and
+          excluding rules.
+        </Text>
+        <Text>
+          Each release of RHEL is supported with a unique and specific version
+          of the SCAP Security Guide (SSG). You must customize each version of
+          SSG for each release of RHEL.
+        </Text>
+      </TextContent>
+      <StateViewWithError
+        stateValues={{
+          error: preselectedRuleIdsError,
+          data: profilesAndRuleIds,
+          loading: preselectedRuleIdsLoading,
+          noRuleSets,
+        }}
+      >
+        <StateViewPart stateKey="noRuleSets">
+          <EmptyState>
+            <EmptyStateHeader
+              titleText="No rules can be configured"
+              headingLevel="h1"
+            />
+            <EmptyStateBody>
+              The policy type selected does not exist for the systems and OS
+              versions selected in the previous steps.
+            </EmptyStateBody>
+          </EmptyState>
+        </StateViewPart>
+        <StateViewPart stateKey="loading">
+          <EmptyState>
+            <Spinner />
+          </EmptyState>
+        </StateViewPart>
+        <StateViewPart stateKey="data">
+          <Tailorings
+            profiles={profilesAndRuleIds}
+            columns={[Columns.Name, Columns.Severity, Columns.Remediation]}
+            ouiaId="RHELVersions"
+            onSelect={onSelect}
+            preselected={preselected}
+            enableSecurityGuideRulesToggle
+            rulesPageLink={true}
+            valueOverrides={valueOverrides}
+            onValueOverrideSave={onValueOverrideSave}
+            selectedVersionCounts={osMinorVersionCounts.reduce(
+              (prev, cur) => ({
+                ...prev,
+                [cur.osMinorVersion]: cur.count,
+              }),
+              {}
+            )}
+          />
+        </StateViewPart>
+      </StateViewWithError>
+    </React.Fragment>
   );
 };
 
 EditPolicyProfilesRules.propTypes = {
-  policy: propTypes.object,
+  profile: propTypes.object,
   change: reduxFormPropTypes.change,
   osMajorVersion: propTypes.string,
   osMinorVersionCounts: propTypes.arrayOf(
@@ -38,7 +219,6 @@ EditPolicyProfilesRules.propTypes = {
     })
   ),
   selectedRuleRefIds: propTypes.array,
-  ruleValues: propTypes.array,
   valueOverrides: propTypes.object,
 };
 
@@ -46,11 +226,10 @@ const selector = formValueSelector('policyForm');
 
 export default compose(
   connect((state) => ({
-    policy: selector(state, 'profile'), // TODO: use "profile" naming instead of "policy"
+    profile: selector(state, 'profile'),
     osMajorVersion: selector(state, 'osMajorVersion'),
     osMinorVersionCounts: selector(state, 'osMinorVersionCounts'),
     selectedRuleRefIds: selector(state, 'selectedRuleRefIds'),
-    ruleValues: selector(state, 'ruleValues'),
     valueOverrides: selector(state, 'valueOverrides'),
   })),
   reduxForm({
